previous_sprints/etl/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
backoff==2.2.1
pydantic==1.10.12